Separate Queue Controller in a different daemon, issue 233 with fix to test code.
