+++ pkg/batchd/cache/cache_test.go
@@ -28,7 +28,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/intstr"
 
 	"github.com/kubernetes-incubator/kube-arbitrator/pkg/batchd/api"
-	arbv1 "github.com/kubernetes-incubator/kube-arbitrator/pkg/batchd/apis/v1"
+	arbv1 "github.com/kubernetes-incubator/kube-arbitrator/pkg/queue-ctrl/apis/v1"
 )
 
 func nodesEqual(l, r map[string]*api.NodeInfo) bool {
